Make rates showing tokens and improve architecture
