+++ src/components/SettingsPage/index.js
@@ -11,7 +11,7 @@ import { accountsSelector } from 'reducers/accounts'
 import Pills from 'components/base/Pills'
 import Box from 'components/base/Box'
 import SectionDisplay from './sections/General'
-import SectionCurrencies from './sections/Currencies'
+import SectionCryptoAssets from './sections/CryptoAssets'
 import SectionHelp from './sections/Help'
 import SectionAbout from './sections/About'
 import SectionExperimental from './sections/Experimental'
@@ -46,7 +46,7 @@ class SettingsPage extends PureComponent<Props, State> {
       {
         key: 'currencies',
         label: <Trans i18nKey="settings.tabs.cryptoAssets" />,
-        value: SectionCurrencies,
+        value: SectionCryptoAssets,
       },
       {
         key: 'export',
